requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
celery
redis  # If you decide to use Redis as broker/backend later
sse-starlette
nltk
psutil
pydantic-settings